Add a function to get the plays for a given user
This exists in the private schema, since it will be run by cron. It
takes in a user ID, and uses their refresh token (if possible) to query
Spotify for their top 50 songs. The "after" pointer is saved, so that
only new tracks will be received by the next pass