Remove all Firebase infrastructure
This removes all references to Firebase from the repo, including in
GitHub Actions scripts. Those will be updated at a later date to work
with Vercel/Supabase. The repo structure no longer needs to be a
monorepo (since there is only one package now), and there are some
duplicate files (i.e., the DB schema TS) but that will be fixed at a
later date