Correctly subscribe to pairing codes on signup
Previously, there was a bug where the generic anon token would be used
to try and subscribe to a pairing code channel. This obviously wouldn't
work. The fix seems to be to force-update the auth state (via
supabase.realtime.setAuth()). There doesn't seem to be a GitHub Issue
related to this, and the obvious check of "ensure there is a user before
listening" doesn't seem to address this. At any rate, this is the
easiest solution and doesn't introduce any additional complexity. This
also renames the fields passed by the server loader to provide some more
clarity.