Read user plays on refresh token update
This ensures that a user who signs in for the first time (or for the
first time in a while) will see updated songs. This will only occur if
the last read is more than 15 minutes old. Checking that condition
requires storing the last read date as a part of the play metadata