Better data loading for unauthenticated users
Previously, unauthenticated users would see a warning that their data
couldn't be loaded (which makes no sense). Also, if you go through the
login flow for the first time, bu the time you got to the dashboard you
would always see the "data load complete". Now, the toast will only
show if you're on a page that needs it when you need it